refactor: use existing information to calculate chars to avoid additional argument passing
