+++ lua/fastaction/window.lua
@@ -1,4 +1,5 @@
 local M = {}
+local keys = require 'fastaction.keys'
 local m = {}
 
 m.namespace = vim.api.nvim_create_namespace 'windmenu'
@@ -107,8 +108,9 @@ function M.popup_window(content, on_buf_create, opts)
     vim.api.nvim_buf_add_highlight(buffer, m.namespace, opts.highlight.divider, 1, 0, -1)
 
     local line = 2 -- avoid the title and the divider i.e. start at line 2
+    local chars = math.floor((#content - 2) / (26 - #keys.filter_alpha_keys(opts.dismiss_keys)))
     for _, _ in pairs(content) do
-        vim.api.nvim_buf_add_highlight(buffer, m.namespace, 'MoreMsg', line, 0, 3 + opts.chars)
+        vim.api.nvim_buf_add_highlight(buffer, m.namespace, 'MoreMsg', line, 0, 3 + chars)
         line = line + 1
     end
     vim.api.nvim_set_option_value('modifiable', false, { buf = buffer })
